requirements.txt
tk
opencv-python
numpy
Pillow
matplotlib
